There are also currently some problems shown by no-segfault.R, some of
which also appear to afflict trunk.


git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/branches/provenance@1336 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/library/Recommended/MASS/src/lqs.c
@@ -64,7 +64,7 @@ static void sample_noreplace(int *x, int n, int k)
 
     for (i = 0; i < n; i++) ind[i] = i;
     for (i = 0; i < k; i++) {
-	j = nn * unif_rand();
+	j = (int)(nn * unif_rand());
 	x[i] = ind[j];
 	ind[j] = ind[--nn];
     }
